Backend/requirements.txt
# Complete Candidate Portal System Dependencies

# Core Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0

# Database
mysql-connector-python>=8.0.0

# Authentication and Security
PyJWT>=2.8.0
bcrypt>=4.0.0

# Resume Filtering System Dependencies
# Core ML and Data Processing
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0

# NLP and Text Processing
spacy>=3.0.0
fuzzywuzzy>=0.18.0
jellyfish>=0.9.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Image Processing (for CAPTCHA)
Pillow>=9.0.0

# Additional Utilities
python-Levenshtein>=0.12.0  # Optional: speeds up fuzzywuzzy
Werkzeug>=2.3.0

# Email Support
secure-smtplib

# Development and Testing
python-dotenv>=1.0.0

# AI Chatbot Dependencies
autogen>=0.2.0
openai>=1.0.0

# Google Meet API Dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0